added usage search for building types

+++ sophena/src/sophena/rcp/editors/basedata/buildings/BuildingStateEditor.java
@@ -14,12 +14,15 @@ import org.eclipse.ui.forms.widgets.ScrolledForm;
 import org.eclipse.ui.forms.widgets.Section;
 
 import sophena.db.daos.RootEntityDao;
+import sophena.db.usage.SearchResult;
+import sophena.db.usage.UsageSearch;
 import sophena.model.BuildingState;
 import sophena.model.BuildingType;
 import sophena.rcp.App;
 import sophena.rcp.Icon;
 import sophena.rcp.M;
 import sophena.rcp.editors.Editor;
+import sophena.rcp.editors.basedata.UsageError;
 import sophena.rcp.utils.Actions;
 import sophena.rcp.utils.Editors;
 import sophena.rcp.utils.KeyEditorInput;
@@ -138,7 +141,11 @@ public class BuildingStateEditor extends Editor {
 					"Soll der ausgewählte Gebäudezustand wirklich gelöscht werden?");
 			if (!doIt)
 				return;
-			// TODO: usage search !!!
+			List<SearchResult> usage = new UsageSearch(App.getDb()).of(s);
+			if (!usage.isEmpty()) {
+				UsageError.show(usage);
+				return;
+			}
 			dao.delete(s);
 			states.remove(s);
 			table.setInput(states);
